+++ temp_cata/modules/nameFix.lua
@@ -225,6 +225,9 @@ local function PartyFrameNameChange(frame)
         frame.bbfName:SetText("")
         return
     end
+    frame.bbfName:ClearAllPoints()
+    frame.bbfName:SetPoint("LEFT", frame.name, "LEFT")
+    frame.bbfName:SetWidth(frame.name:GetWidth())
     if partyArenaNames and IsActiveBattlefieldArena() then
         SetArenaName(frame, frame.unit, frame.bbfName)
         return
@@ -319,6 +322,7 @@ local function UpdateFontStringPosition(frame)
     if point then
         frame.bbfName:ClearAllPoints()
         frame.bbfName:SetPoint(point, relativeTo, relativePoint, xOffset, yOffset)
+        frame.bbfName:SetWidth(name:GetWidth())
     end
 end